# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: 豆芽成长助手 API
  version: 0.1.0
  description: |
    托管班系统 API 契约。当前已接入基础档案中心、身份权限、接送、请假、作业和家长端闭环。
servers:
  - url: http://localhost:8080
    description: Local development

# 业务接口使用 bearerAuth；登录和微信 code2Session 入口不需要 bearerAuth。
security: [{ bearerAuth: [] }]
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: 管理员或教师登录
      operationId: loginStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, minLength: 3 }
                password: { type: string, minLength: 6 }
              additionalProperties: false
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: 刷新登录 Token
      operationId: refreshAuthToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
              additionalProperties: false
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 当前登录身份
      operationId: getAuthMe
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 当前身份
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthMeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/codes:
    get:
      tags: [Auth]
      summary: 当前身份权限码
      operationId: getAuthCodes
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 权限码列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties: { data: { type: array, items: { type: string } } }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录
      operationId: logoutAuth
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 退出成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties: { data: { type: boolean } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-hmac-token
...
  headers:
    RequestID:
      description: 当前请求 ID。
      schema:
        type: string
      example: req-123.test
  responses:
    BadRequest:
      description: 请求不合法
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 10001
            message: 请求不合法
    Unauthorized:
      description: 未登录或登录状态已失效
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 20001
            message: 认证失败
...
  schemas:
    AuthTokenResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required: [accessToken, refreshToken, expiresIn, principal, role]
              properties:
                accessToken: { type: string }
                refreshToken: { type: string }
                expiresIn: { type: integer, format: int64 }
                principal: { type: string, enum: [user, parent] }
                role: { type: string }
              additionalProperties: false
...
    AuthMeResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required: [id, realName, role, principal, homePath]
              properties:
                id: { type: integer, format: int64 }
                username: { type: string }
                realName: { type: string }
                nickname: { type: string }
                avatar: { type: string }
                role: { type: string }
                roles: { type: array, items: { type: string } }
                principal: { type: string, enum: [user, parent] }
                homePath: { type: string }
              additionalProperties: false
    SuccessResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          description: 业务响应码。成功固定为 0。
          example: 0
        message:
          type: string
          example: success
        data:
          description: 响应数据。
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
          description: 业务错误码。
          enum: [10001, 10002, 10003, 10004, 10005, 10006, 10007, 20001, 20003, 50000, 50001]
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationDetail'
      additionalProperties: false
    ValidationDetail:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
          example: username
        reason:
          type: string
          example: 必填
      additionalProperties: false
...